test(integration): collapse the boot suite back to two shards
The first run on a busy macOS pool surfaced the trade-off in
splitting too finely. Only two of the five iOS shards got
runners; the other three sat queueing on "All GitHub-hosted
runners with label [macos-latest] are busy" — GitHub's global
Mac fleet, not our per-account cap. On those evenings the
five-way fan-out reverts to roughly two waves of effective
parallelism, with the queue depth gating wall-clock time
instead of our work doing it.

Group the five files into the two natural questions a smoke
asks of the boot:

- app_boot_test.dart: did main() finish, and did anything fire
  FlutterError.onError along the way? One simulator boot
  answers both — same setup window, two assertions sharing it.
- first_screen_test.dart: what does a fresh-install user
  actually see on page one? Routing chose OnboardingScreen,
  OnboardingBloc reached loaded state so IntroductionScreen
  mounted, the intl delegates loaded so a known ARB string
  rendered. Three layers of the same observation, one boot.

The matrix infrastructure is unchanged — discover-integration-
tests still fans out at file granularity, gated on a clean
build. Two shards is friendlier to the pool today, and the
shape lets us grow back toward more shards organically as more
genuine smoke surfaces appear, without redesigning the workflow.
